Elaborate on expected time to complete the test
We do not require the coding test to be completed in a set amount of time.
To avoid wasting too much of your time though, it is best to not spend more
than a couple of hours on the solution.

A top level README is a good place to elaborate on your thoughts for what your
next steps would be to complete and make a production-ready solution.