Update tech test to utilise jsonapi spec
Feedback from a previous submission was that the purchase route is not
what they expected RESTful wise. Switching the examples to use jsonapi
spec including updating the swagger / postman collection. Swagger was
generated from the postman collection with postman2openapi
(https://kevinswiber.github.io/postman2openapi/).

Updated the `desc` to `description` as this caused confusion with a
previous submission as they assumed that was descending.